pyproject.toml
[project]
name = "fundamental-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pytest>=8.4.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.39.0",
]
